requirements.txt
gensim
requests
html2text

# dev
pylint
pytest>=3.7.3   # https://github.com/pytest-dev/pytest/issues/3819
pytest-cov
python-coveralls
mypy
sphinx
sphinx-autodoc-typehints
